More code added
Now stuck on duplicate zips. which are complicating the join. The crosswalk did its job, but now we  need to create a new frame that has much of the same information, it just needs to combine duplicate rows so we can sum contribtuions and join the shapefile so there is one zip for each shape.